requirements.txt
PyYAML
wordnik
